+++ CHANGELOG.md
@@ -56,6 +56,7 @@
   `PFractional`
 * `Plutarch.Num` is now `Plutarch.Internal.Numeric`
 * `PIntegral` type class is now in `Plutarch.Internal.Numeric`
+* `Plutarch.Integer` is now `Plutarch.Builtin.Integer`
 
 ## Removed